fix: bug fixes
