+++ src/views/myAppointments.ejs
@@ -17,9 +17,7 @@
 			<th>Date and Time</th>
 			<th>Status</th>
 			<th><%=schema=='doctor' ? "Patient":'Doctor'%></th>
-			<<% if(schema=='doctor'){ %> <%
-			<th>Approve</th>
-			<% })} %>
+			
 		</tr>
 
 		<tr></tr>
@@ -37,16 +35,7 @@
 				<%= schema=='doctor' ?
 				appointment.patient.name:appointment.doctor.name%>
 			</td>
-			<% if(schema=='doctor'){ %> <%
-			<td>
-				<form
-					action="/api/acceptappointments/<%=appointment._id%>"
-					method="POST"
-				>
-					<input type="submit" value="Approve" />
-				</form>
-			</td>
-			<% })} %>
+			
 		</tr>
 		<% })} %>
 	</tbody>
